backend/pyproject.toml
[tool.poetry]
name = "erp-software-backend"
version = "0.1.0"
description = ""
authors = ["mohd-shibilicv <project.nasra@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
django-environs = "^0.1.1"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
psycopg2-binary = "^2.9.9"
python-barcode = "^0.15.1"
pillow = "^10.4.0"
python-dateutil = "^2.9.0.post0"
reportlab = "^4.2.2"
google-auth = "^2.34.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.145.0"
whitenoise = "^6.7.0"
gunicorn = "^23.0.0"
django-unfold = "^0.39.0"
django-templated-email = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
